Add commands /autoban and /autokick
The commands are executed sendingin the chat:
    /autokick reason
    /autoban reason

where 'reason' is a message to explain to the user the reason of the kick/ban.

The command must be used in reply to the user's message. The bot send a message in the chat to the user providing the 'reason'. After 60 seconds the user is kicked/banned.

Changes:
    modified:   commands/admin/__init__.py
    new file:   commands/admin/autoban.py
    new file:   commands/admin/autokick.py
    modified:   commands/index.py